Move react-native core into core module
