+++ modules/core/client-react-native/App.tsx
@@ -3,30 +3,33 @@ import { ApolloProvider } from 'react-apollo';
 import { createStore, combineReducers } from 'redux';
 import { Provider } from 'react-redux';
 import url from 'url';
-import log from '../../common/log';
 
-import modules from '../../client/src/modules';
-import createApolloClient from '../../common/createApolloClient';
+import ClientModule from '@module/module-client-react-native';
+import log from '../../../packages/common/log';
+import createApolloClient from '../../../packages/common/createApolloClient';
 
 const { protocol, pathname, port } = url.parse(__API_URL__);
-const store = createStore(
-  combineReducers({
-    ...modules.reducers
-  }),
-  {} // initial state
-);
 
 interface MainProps {
+  children?: any;
   exp: any;
+  modules: ClientModule;
 }
 
 export default class Main extends React.Component<MainProps> {
   public render() {
     const { hostname } = url.parse(__API_URL__);
+    const { modules } = this.props;
     const apiUrl =
       this.props.exp.manifest.bundleUrl && hostname === 'localhost'
         ? `${protocol}//${url.parse(this.props.exp.manifest.bundleUrl).hostname}:${port}${pathname}`
         : __API_URL__;
+    const store = createStore(
+      combineReducers({
+        ...modules.reducers
+      }),
+      {} // initial state
+    );
     const client = createApolloClient({
       apiUrl,
       createNetLink: modules.createNetLink,
