+++ modules/core/client-react-native/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@module/core-client-react-native",
+  "version": "1.0.0",
+  "private": true,
+  "dependencies": {
+    "@module/core-common": "^1.0.0"
+  }
+}
