+++ modules/module/client-react-native/index.ts
@@ -1,0 +1,2 @@
+export { default } from './ClientModule';
+export * from './ClientModule';
