+++ modules/module/client-react-native/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@module/module-client-react-native",
+  "version": "1.0.0",
+  "private": true,
+  "dependencies": {
+    "@module/module-common": "^1.0.0"
+  }
+}
