--- modules/module/client-react/index.native.ts
@@ -1,2 +1,0 @@
-export { default } from './ClientModule.native';
-export * from './ClientModule.native';
