+++ package.json
@@ -68,6 +68,7 @@
   "workspaces": [
     "packages/*",
     "modules/**/client-react",
+    "modules/**/client-react-native",
     "modules/**/client-angular",
     "modules/**/common",
     "modules/**/common-react",
