+++ packages/client/src/modules/ClientModule.native.ts
@@ -1,3 +1,3 @@
 // Reexport for backward-compatibility purposes
-export { default } from '@module/module-client-react/index.native';
-export * from '@module/module-client-react/index.native';
+export { default } from '@module/module-client-react-native';
+export * from '@module/module-client-react-native';
